Moving file-utils from ashell to src
